press done and then check for the field names being present.
Some new change has prevented the update from happening until done is pressed. Kind of a bummmer because I thought it was cool how it "live" imported but it is ultimately not worth debugging right now.

Something to remember when debugging this - the issue might be specific to the map because the team csv import does still refresh live.

Tangent - would subscribing to all changes through graphql subscriptions be simpler then updating after a mutation in my code? It would require the subscription to be reliable...